+++ api/src/controllers/Users/putUser.js
@@ -1,25 +1,21 @@
 const { User } = require('../../db')
+const { userToUpdateUtil } = require("../../utils/userToUpdateUtil")
 
 const updateUser = async (req, res) => {
 
-    const changedDetail = {}
     const {
         user_id, email, fullName, username, picture, birthday
     } = req.body
 
     const userToUpdate = await User.findOne({where: {user_id: user_id}}) 
     
-    if (userToUpdate) changedDetail.id = userToUpdate.dataValues.id
-    changedDetail.fullName = fullName
-    changedDetail.username = username
-    changedDetail.picture = picture
-    changedDetail.birthday = birthday
-    changedDetail.user_id = user_id
-    changedDetail.email = email
+    userToUpdate ?
+    updatedDetail = userToUpdateUtil(userToUpdate, user_id, email, fullName, username, picture, birthday) :
+    res.status(404).json("User not found :(")
 
     try {
-        const [updatedUser, created] = await User.upsert(changedDetail)
-        res.status(200).json(updatedUser)
+        const [updatedUser, created] = await User.upsert(updatedDetail)
+        res.status(200).json(updatedUser || created)
     } catch(error) {
         res.status(404).json(console.log(error))
     }
